chore(platform): rename err to error
