+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-resolver ="2"
+resolver = "2"
 members = [
     "src/Cobalt.Engine/*"
 ]
