+++ src/Cobalt.Engine/platform/src/error.rs
